+++ ft_printf/libft/libft.h
@@ -74,7 +74,7 @@ int					ft_strnequ(char const *s1, char const *s2, size_t n);
 char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
-char				*ft_itoa(int n);
+char				*ft_itoa(long long n);
 char				**ft_strsplit(char const *s, char c);
 void				ft_strdel(char **as);
 void				ft_putendl_fd(char const *s, int fd);
